+++ internal/server/server.go
@@ -40,7 +40,7 @@ func New(ctx context.Context, name, version string, client *pipedrive.Client, do
 	tools.RegisterPipelines(srv, client, domain)
 	tools.RegisterDeals(srv, client, domain, dryRun)
 	tools.RegisterPersons(srv, client, domain, dryRun)
-	tools.RegisterOrganizations(srv, client, domain)
+	tools.RegisterOrganizations(srv, client, domain, dryRun)
 	tools.RegisterActivities(srv, client, domain)
 	tools.RegisterNotes(srv, client, dryRun, enableDestructive)
 	tools.RegisterCache(srv, client)